New frontend (#2)
* frontend basic

* small changes in backend

* connected results and summary

* modified how clients are destroyed

* added markdown support

* deleted unnecessary route

* testing build version

* things

* deleted compile folder

* minor changes

* merged with frontend

* created description for stage wise testing

* execution bar

* verified working status

* added sidebar

* fixed sidebar and added delete button

* minor sidebar changes

* minor sidebar changes

* execute component modification

---------

+++ frontend/src/lib/rest.ts
@@ -13,6 +13,8 @@ export const getToken = async () => {
 }
 
 export const getStageDescription = async (stageNo: number, token: string) => {
+    if (stageNo < 1 || !token)
+        return;
     try {
         const response = await axios.get(`${BACKEND_URL}/stage/${stageNo}`, {
             headers: {
@@ -21,12 +23,15 @@ export const getStageDescription = async (stageNo: number, token: string) => {
         });
         return response.data;
     }
-    catch {
+    catch (error) {
+        console.log('error', error, stageNo);
         return null;
     }
 }
 
 export const uploadBinary = async (stageNo: number, token: string, file: File) => {
+    if (stageNo < 1 || !token)
+        return;
     const formData = new FormData();
     formData.append('binary', file);
 
@@ -36,7 +41,7 @@ export const uploadBinary = async (stageNo: number, token: string, file: File) =
             formData,
             {
                 headers: {
-                    "Content-Type": 'multipart/f',
+                    "Content-Type": 'multipart/form-data',
                     Authorization: `Bearer ${token}`,
                 }
             }
@@ -49,6 +54,8 @@ export const uploadBinary = async (stageNo: number, token: string, file: File) =
 }
 
 export const deleteBinary = async (stageNo: number, token: string) => {
+    if (!token)
+        return;
     try {
         await axios.delete(
             `${BACKEND_URL}/stage/${stageNo}/binary`,
